app/requirements.txt
Flask==2.3.2
redis==4.5.5
psycopg2-binary==2.9.10
gunicorn==20.1.0
